apps/worker/pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "face-ai-worker"
version = "0.1.0"
description = "Face-AI Celery worker — AI processing pipeline"
requires-python = ">=3.11"
dependencies = [
    "celery[redis]>=5.3",
    "onnxruntime>=1.17",
    "insightface>=0.7.3",
    "open_clip_torch>=2.24",
    "torch>=2.2",
    "torchvision>=0.17",
    "Pillow>=10.0",
    "piexif>=1.1",
    "numpy>=1.26",
    "opencv-python-headless>=4.9",
    "boto3>=1.34",
    "qdrant-client>=1.8",
    "pydantic-settings>=2.0",
    "prisma>=0.15",
]
